Discover SDP data by downloading catalog information.

Discover SDP data by downloading catalog information.

Usage

sdp_get_catalog(
  domains = c("UG", "UER", "GT"),
  types = c("Mask", "Topo", "Vegetation", "Hydro", "Planning", "Radiation", "Snow",
    "Climate", "Imagery", "Supplemental"),
  releases = c("Basemaps", "Release1", "Release2", "Release3", "Release4"),
  timeseries_types = c("Single", "Yearly", "Seasonal", "Monthly", "Daily"),
  deprecated = FALSE,
  return_stac = FALSE
)

Arguments

domains

The spatial domain of the desired data product.

types

The type of product to return.

releases

Which release (group) of data products to return.

timeseries_types

Some datasets are structured as single datasets (e.g. Single), and others are time-series with various periods (e.g.Monthly).

deprecated

Should older versions of datasets be returned, or just the latest version?

return_stac

NOT IMPLEMENTED: Should the results be returned as a Spatio-temporal Asset Catalog? Otherwise return and ordinary data frame.

Value

A data frame containing basic catalog information for the matched data products.

Examples

## Gets information on all current datasets.
sdp_cat <- sdp_get_catalog()
str(sdp_cat)
#> 'data.frame':	78 obs. of  17 variables:
#>  $ CatalogID      : chr  "R1D001" "R1D002" "R1D003" "R1D004" ...
#>  $ Release        : chr  "Release1" "Release1" "Release1" "Release1" ...
#>  $ Type           : chr  "Hydro" "Hydro" "Hydro" "Hydro" ...
#>  $ Product        : chr  "Large Stream Flowlines (Multi-flow Direction)" "Large Stream Flowlines (Single-flow Direction)" "Large Watersheds (Multi-flow Direction)" "Multi-direction Flow Accumulation" ...
#>  $ Domain         : chr  "UER" "UER" "UER" "UER" ...
#>  $ Resolution     : chr  "1m" "1m" "1m" "1m" ...
#>  $ Deprecated     : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ MinDate        : chr  "7/16/18" "7/16/18" "7/16/18" "7/16/18" ...
#>  $ MaxDate        : chr  "7/16/18" "7/16/18" "7/16/18" "7/16/18" ...
#>  $ MinYear        : int  2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 ...
#>  $ MaxYear        : int  2018 2018 2018 2018 2018 2018 2018 2018 2018 2018 ...
#>  $ TimeSeriesType : chr  "Single" "Single" "Single" "Single" ...
#>  $ DataUnit       : chr  "categorical" "categorical" "categorical" "square meters" ...
#>  $ DataScaleFactor: int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ DataOffset     : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ Data.URL       : chr  "https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release1/UER_streams_512k_mfd_1m_v2.tif" "https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release1/UER_streams_512k_sfd_1m_v2.tif" "https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release1/UER_basins_512k_mfd_1m_v2.tif" "https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release1/UER_flow_accum_mfd_1m_v3.tif" ...
#>  $ Metadata.URL   : chr  "https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release1/UER_streams_512k_mfd_1m_v2_metadata.xml" "https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release1/UER_streams_512k_sfd_1m_v2_metadata.xml" "https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release1/UER_basins_512k_mfd_1m_v2_metadata.xml" "https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release1/UER_flow_accum_mfd_1m_v3_metadata.xml" ...

## Gets a subset of catalog entries for the Upper Gunnison (UG) domain.
sdp_sub <- sdp_get_catalog(domains="UG", types="Vegetation",
                          deprecated=FALSE,return_stac=FALSE)
sdp_sub
#>    CatalogID  Release       Type                       Product Domain
#> 54    R3D013 Release3 Vegetation              Understory Cover     UG
#> 55    R3D014 Release3 Vegetation       Vegetation Canopy Cover     UG
#> 56    R3D015 Release3 Vegetation      Vegetation Canopy Height     UG
#> 57    R3D016 Release3 Vegetation 20th Percentile Canopy Height     UG
#> 58    R3D017 Release3 Vegetation 80th Percentile Canopy Height     UG
#> 59    R3D018 Release3 Vegetation               Basic Landcover     UG
#> 60    R3D019 Release3 Vegetation        October 2017 NAIP NDVI     UG
#> 61    R3D020 Release3 Vegetation       Septober 2019 NAIP NDVI     UG
#> 73     BM012 Basemaps Vegetation      Canopy Structure Basemap     UG
#> 74     BM013 Basemaps Vegetation             Landcover Basemap     UG
#>    Resolution Deprecated MinDate MaxDate MinYear MaxYear TimeSeriesType
#> 54         3m      FALSE  9/7/19  9/7/19    2019    2019         Single
#> 55         3m      FALSE  9/7/19  9/7/19    2019    2019         Single
#> 56         1m      FALSE  9/7/19  9/7/19    2019    2019         Single
#> 57         3m      FALSE  9/7/19  9/7/19    2019    2019         Single
#> 58         3m      FALSE  9/7/19  9/7/19    2019    2019         Single
#> 59         1m      FALSE  9/7/19  9/7/19    2019    2019         Single
#> 60         1m      FALSE  9/7/19  9/7/19    2019    2019         Single
#> 61         1m      FALSE  9/7/19  9/7/19    2019    2019         Single
#> 73         2m      FALSE  9/7/19  9/7/19    2019    2019         Single
#> 74         2m      FALSE  9/7/19  9/7/19    2019    2019         Single
#>       DataUnit DataScaleFactor DataOffset
#> 54  proportion               1          0
#> 55  proportion               1          0
#> 56      meters               1          0
#> 57      meters               1          0
#> 58      meters               1          0
#> 59 categorical               1          0
#> 60    unitless               1          0
#> 61    unitless               1          0
#> 73    unitless               1          0
#> 74    unitless               1          0
#>                                                                                                   Data.URL
#> 54 https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_canopy_2mcover_3m_v2.tif
#> 55   https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_canopy_cover_3m_v3.tif
#> 56  https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_canopy_height_1m_v2.tif
#> 57    https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_canopy_ht20_3m_v2.tif
#> 58    https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_canopy_ht80_3m_v4.tif
#> 59      https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_landcover_1m_v4.tif
#> 60   https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_ndvi_oct2017_1m_v1.tif
#> 61  https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_ndvi_sept2019_1m_v2.tif
#> 73             https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/basemaps/UG_canopy_basemap_v3.tif
#> 74          https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/basemaps/UG_landcover_basemap_v3.tif
#>                                                                                                        Metadata.URL
#> 54 https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_canopy_2mcover_3m_v2_metadata.xml
#> 55   https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_canopy_cover_3m_v3_metadata.xml
#> 56  https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_canopy_height_1m_v2_metadata.xml
#> 57    https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_canopy_ht20_3m_v2_metadata.xml
#> 58    https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_canopy_ht80_3m_v4_metadata.xml
#> 59      https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_landcover_1m_v4_metadata.xml
#> 60   https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_ndvi_oct2017_1m_v1_metadata.xml
#> 61  https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release3/UG_ndvi_sept2019_1m_v2_metadata.xml
#> 73             https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/basemaps/UG_canopy_basemap_v3_metadata.xml
#> 74          https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/basemaps/UG_landcover_basemap_v3_metadata.xml