Download detailed geospatial metadata for SDP Datasets.

Download detailed geospatial metadata for SDP Datasets.

Usage

sdp_get_metadata(catalog_id, return_list = TRUE)

Arguments

catalog_id

The unique Catalog ID code for the desired dataset.

return_list

Logical. If TRUE, then the output is parsed as an R list object.

Value

A nested list (return_list=TRUE) or XML document (return_list=FALSE) containing detailed geospatial metadata for each dataset.

Examples

##Get metadata for a specific item.
sdp_get_metadata(catalog_id="R1D001",return_list=TRUE)
#> $qgis
#> $qgis$identifier
#> $qgis$identifier[[1]]
#> [1] "/vsicurl/https://rmbl-sdp.s3.us-east-2.amazonaws.com/data_products/released/release1/UER_streams_512k_mfd_1m_v2.tif"
#> 
#> 
#> $qgis$parentidentifier
#> list()
#> 
#> $qgis$language
#> $qgis$language[[1]]
#> [1] "en"
#> 
#> 
#> $qgis$type
#> $qgis$type[[1]]
#> [1] "dataset"
#> 
#> 
#> $qgis$title
#> $qgis$title[[1]]
#> [1] "Map of Multi-direction Stream Flowlines from the 2018 NEON AOP Digital Elevation Model"
#> 
#> 
#> $qgis$abstract
#> $qgis$abstract[[1]]
#> [1] "This map represents estimated stream flowlines from a hydrologically corrected digital elevation model. The lines were derived in GRASS GIS using a multi-direction algorithm that allows channel braiding. Each stream segment is identified by a unique integer. Stream lines were delineated for drainage areas greater than 512000 square meters.\n"
#> 
#> 
#> $qgis$keywords
#> $qgis$keywords$keyword
#> $qgis$keywords$keyword[[1]]
#> [1] "Environment"
#> 
#> 
#> $qgis$keywords$keyword
#> $qgis$keywords$keyword[[1]]
#> [1] "Geoscientific Information"
#> 
#> 
#> $qgis$keywords$keyword
#> $qgis$keywords$keyword[[1]]
#> [1] "Inland Waters"
#> 
#> 
#> $qgis$keywords$keyword
#> $qgis$keywords$keyword[[1]]
#> [1] "Society"
#> 
#> 
#> attr(,"vocabulary")
#> [1] "gmd:topicCategory"
#> 
#> $qgis$contact
#> $qgis$contact$name
#> $qgis$contact$name[[1]]
#> [1] "Ian Breckheimer"
#> 
#> 
#> $qgis$contact$organization
#> $qgis$contact$organization[[1]]
#> [1] "Rocky Mountain Biological Laboratory"
#> 
#> 
#> $qgis$contact$position
#> $qgis$contact$position[[1]]
#> [1] "Research Scientist"
#> 
#> 
#> $qgis$contact$voice
#> list()
#> 
#> $qgis$contact$fax
#> list()
#> 
#> $qgis$contact$email
#> $qgis$contact$email[[1]]
#> [1] "ikb@rmbl.org"
#> 
#> 
#> $qgis$contact$role
#> $qgis$contact$role[[1]]
#> [1] "owner"
#> 
#> 
#> 
#> $qgis$links
#> $qgis$links$link
#> list()
#> attr(,"url")
#> [1] "https://rmbl.org/data"
#> attr(,"name")
#> [1] "URL"
#> attr(,"format")
#> [1] ""
#> attr(,"size")
#> [1] ""
#> attr(,"type")
#> [1] "WWW:LINK"
#> attr(,"mimeType")
#> [1] ""
#> attr(,"description")
#> [1] "Spatial Data Platform Description"
#> 
#> 
#> $qgis$fees
#> list()
#> 
#> $qgis$license
#> $qgis$license[[1]]
#> [1] "Creative Commons Attribution 4.0"
#> 
#> 
#> $qgis$encoding
#> list()
#> 
#> $qgis$crs
#> $qgis$crs$spatialrefsys
#> $qgis$crs$spatialrefsys$wkt
#> $qgis$crs$spatialrefsys$wkt[[1]]
#> [1] "PROJCRS[\"WGS 84 / UTM zone 13N\",BASEGEOGCRS[\"WGS 84\",DATUM[\"World Geodetic System 1984\",ELLIPSOID[\"WGS 84\",6378137,298.257223563,LENGTHUNIT[\"metre\",1]]],PRIMEM[\"Greenwich\",0,ANGLEUNIT[\"degree\",0.0174532925199433]],ID[\"EPSG\",4326]],CONVERSION[\"UTM zone 13N\",METHOD[\"Transverse Mercator\",ID[\"EPSG\",9807]],PARAMETER[\"Latitude of natural origin\",0,ANGLEUNIT[\"degree\",0.0174532925199433],ID[\"EPSG\",8801]],PARAMETER[\"Longitude of natural origin\",-105,ANGLEUNIT[\"degree\",0.0174532925199433],ID[\"EPSG\",8802]],PARAMETER[\"Scale factor at natural origin\",0.9996,SCALEUNIT[\"unity\",1],ID[\"EPSG\",8805]],PARAMETER[\"False easting\",500000,LENGTHUNIT[\"metre\",1],ID[\"EPSG\",8806]],PARAMETER[\"False northing\",0,LENGTHUNIT[\"metre\",1],ID[\"EPSG\",8807]]],CS[Cartesian,2],AXIS[\"(E)\",east,ORDER[1],LENGTHUNIT[\"metre\",1]],AXIS[\"(N)\",north,ORDER[2],LENGTHUNIT[\"metre\",1]],USAGE[SCOPE[\"unknown\"],AREA[\"World - N hemisphere - 108°W to 102°W - by country\"],BBOX[0,-108,84,-102]],ID[\"EPSG\",32613]]"
#> 
#> 
#> $qgis$crs$spatialrefsys$proj4
#> $qgis$crs$spatialrefsys$proj4[[1]]
#> [1] "+proj=utm +zone=13 +datum=WGS84 +units=m +no_defs"
#> 
#> 
#> $qgis$crs$spatialrefsys$srsid
#> $qgis$crs$spatialrefsys$srsid[[1]]
#> [1] "3097"
#> 
#> 
#> $qgis$crs$spatialrefsys$srid
#> $qgis$crs$spatialrefsys$srid[[1]]
#> [1] "32613"
#> 
#> 
#> $qgis$crs$spatialrefsys$authid
#> $qgis$crs$spatialrefsys$authid[[1]]
#> [1] "EPSG:32613"
#> 
#> 
#> $qgis$crs$spatialrefsys$description
#> $qgis$crs$spatialrefsys$description[[1]]
#> [1] "WGS 84 / UTM zone 13N"
#> 
#> 
#> $qgis$crs$spatialrefsys$projectionacronym
#> $qgis$crs$spatialrefsys$projectionacronym[[1]]
#> [1] "utm"
#> 
#> 
#> $qgis$crs$spatialrefsys$ellipsoidacronym
#> $qgis$crs$spatialrefsys$ellipsoidacronym[[1]]
#> [1] "EPSG:7030"
#> 
#> 
#> $qgis$crs$spatialrefsys$geographicflag
#> $qgis$crs$spatialrefsys$geographicflag[[1]]
#> [1] "false"
#> 
#> 
#> 
#> 
#> $qgis$extent
#> $qgis$extent$spatial
#> list()
#> attr(,"dimensions")
#> [1] "2"
#> attr(,"minx")
#> [1] "315130"
#> attr(,"minz")
#> [1] "0"
#> attr(,"maxx")
#> [1] "337845"
#> attr(,"maxz")
#> [1] "0"
#> attr(,"miny")
#> [1] "4297480"
#> attr(,"maxy")
#> [1] "4323209"
#> attr(,"crs")
#> [1] "EPSG:32613"
#> 
#> $qgis$extent$temporal
#> $qgis$extent$temporal$period
#> $qgis$extent$temporal$period$start
#> list()
#> 
#> $qgis$extent$temporal$period$end
#> list()
#> 
#> 
#> 
#> 
#> attr(,"version")
#> [1] "3.10.2-A Coruña"
#>